requirements.txt
numpy==2.4.3
pandas==2.3.3
streamlit==1.55.0
scikit-learn==1.8.0
Pillow==12.1.1
click==8.3.1
protobuf==6.33.5